% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sasformat.R
\name{check_sasformat}
\alias{check_sasformat}
\title{Check a SASformat object}
\usage{
check_sasformat(sasformat)
}
\arguments{
\item{sasformat}{a `SASformat` object}
}
\value{
nothing, errors as a side effect
}
\description{
This is used to ensure that the correct formats and catalogs exist at runtime.
}
