% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format.SASformat}
\alias{format.SASformat}
\title{S3 method for class 'SASformat'}
\usage{
\method{format}{SASformat}(x, ...)
}
\arguments{
\item{x}{A vector or list of type SASformat}

\item{...}{Not used}
}
\value{
A character vector of a formatted SASformat object
}
\description{
S3 method for class 'SASformat'
}
\examples{
proc_format(
  AGEGR1 = value(
    "< 21" = "<21",
    "21 - 39" = "21 - 39",
    "40 <=" = ">=40"
  )
)
age <- SASformat(c(10,20,30,40,50),
                 "AGEGR1")

format(age)
}
