% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sasformat.R
\name{init_sasformat}
\alias{init_sasformat}
\title{Internally create a `SASformat` object}
\usage{
init_sasformat(vector, format, catalog)
}
\arguments{
\item{vector}{see `SASformat()`}

\item{format}{see `SASformat()`}

\item{catalog}{see `SASformat()`}
}
\value{
An object of type `SASformat`
}
\description{
Internally create a `SASformat` object
}
