% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.SASformat}
\alias{print.SASformat}
\title{S3 method for class 'SASformat'}
\usage{
\method{print}{SASformat}(x, ...)
}
\arguments{
\item{x}{A}

\item{...}{Not used}
}
\description{
S3 method for class 'SASformat'
}
\examples{
proc_format(
  AGEGR1 = value(
    "< 21" = "<21",
    "21 - 39" = "21 - 39",
    "40 <=" = ">=40"
  )
)
age <- SASformat(c(10,20,30,40,50),
                 "AGEGR1")

print(age)
}
