% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_format.R
\name{proc_format}
\alias{proc_format}
\title{Create a format catalog}
\usage{
proc_format(..., catalog = "formats")
}
\arguments{
\item{...}{See `vignette("getting_started", package = "SASformatR")`}

\item{catalog}{(`character`) Name of object to store format catalog in}
}
\value{
nothing, creates an object `formats` as a side effect
}
\description{
Create a format catalog
}
\examples{
proc_format(
AGEGR1 = value(
  "< 21" = "<21",
  "21 - 39" = "21 - 39",
  "40 <=" = ">=40"
)
)
}
