% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put.R
\name{put}
\alias{put}
\alias{input}
\title{Recode values based off a format in a catalog}
\usage{
put(x, format, catalog = "formats")

input(x, format, catalog = "formats")
}
\arguments{
\item{x}{A vector or list to apply the format to.}

\item{format}{(`character`) The name of the format to apply}

\item{catalog}{(`character`) The name of the catalog to get the format from - typically `formatsß`}
}
\value{
A vector or list of the type indicated by the format.
}
\description{
Recode values based off a format in a catalog
}
\examples{
proc_format(
 AGEGR1 = value(
   "< 21" = "<21",
   "21 - 39" = "21 - 39",
   "40 <=" = ">=40"
 )
)
age <- c(10,20,30,40,50)
agegr1 <- put(age, "AGEGR1")
}
