% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sasformat.R
\name{SASformat}
\alias{SASformat}
\title{Create a `SASformat` object}
\usage{
SASformat(vector, format, catalog = "formats")
}
\arguments{
\item{vector}{Any list or vector}

\item{format}{(`character`) Name of format to apply}

\item{catalog}{(`character`) Name of catalog to find format in}
}
\value{
A `SASformat` vector, with appropriate methods
}
\description{
Create a `SASformat` object
}
\examples{
proc_format(
 AGEGR1 = value(
   "< 21" = "<21",
   "21 - 39" = "21 - 39",
   "40 <=" = ">=40"
. )
)
age <- c(10,20,30,40,50)
agegr1 <- SASformat(age, "AGEGR1")
}
