% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{[.SASformat}
\alias{[.SASformat}
\title{S3 method for 'SASformat'}
\usage{
\method{`[`}{SASformat}(x, ...)
}
\arguments{
\item{x}{A `SASformat` object}

\item{...}{Passed to the base method}
}
\value{
An subset `SASformat` object
}
\description{
S3 method for 'SASformat'
}
\examples{
proc_format(
  AGEGR1 = value(
    "< 21" = "<21",
    "21 - 39" = "21 - 39",
    "40 <=" = ">=40"
  )
)
age <- SASformat(c(10,20,30,40,50),
                 "AGEGR1")

age[1]
}
